// src/shared/types.ts
export type RoleName = string;

export interface Role {
  name: string;
  description?: string;
  permissions: Permission[];
  isSystem?: boolean; // Se true, não pode ser excluído
  userCount?: number; // Campo calculado para listagem
}
// ...
export type Permission = 
  | 'PENDENCIA:LER_TODAS'
  | 'PENDENCIA:CRIAR'
  | 'PENDENCIA:EDITAR' // Edição geral (título, descrição, prioridade)
  | 'PENDENCIA:MOVER'  // Mudança de status via Kanban
  | 'PENDENCIA:CONCLUIR'
  | 'PENDENCIA:CANCELAR'
  | 'PENDENCIA:ATRIBUIR_RESPONSAVEL'
  | 'USUARIO:GERENCIAR'
  | 'OS:CRIAR'
  | 'SISTEMA:CONFIGURAR';